src/bridge/requirements.txt
boto3>=1.28.0
botocore>=1.31.0
google-cloud-pubsub>=2.18.0
python-dotenv>=1.0.0
